modified code to use static_cast<> to convert to double properly.

+++ Activity8CodeCorrection.cpp
@@ -4,13 +4,12 @@ using namespace std;
 int main() {
 
     double x = 10;
-
     int y = 3;
 
-    double z = x / y;
+    // Use static_cast to convert y to double before performing the division
+    double z = x / static_cast<double>(y);
 
-    cout << z;
+    cout << z;  // This will print the result of the division as a double
 
     return 0;
-
 }
